updated display

+++ blogPosts/views.py
@@ -34,7 +34,7 @@ def post_create(request):
 
 def post_detail(request, slug=None):
     instance = get_object_or_404(Post, slug=slug)
-    
+
     share_string = quote_plus(instance.content)
 
     initial_data = {
@@ -138,6 +138,8 @@ def post_update(request, slug=None):
 
 
 def post_delete(request, slug=None):
+    if not request.user.is_authenticated():
+        raise Http404
     instance = get_object_or_404(Post, slug=slug)
     instance.delete()
     messages.success(request, "Successfully deleted")
